django_backend/requirements.txt
gunicorn
psycopg2
django
whitenoise
djangorestframework
requests
